Fix: move (App.js) and it's test file to single folder
