+++ src/components/App/App.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
-import withRestrictedAccess from './components/hoc/withRestrictedAccess';
-import Dashboard from './components/pages/Dashboard/Dashboard';
-import Login from './components/pages/Login/Login';
-import CreateAccount from './components/pages/CreateAccount/CreateAccount';
-import { getUser } from './utils/handlers/authenticationHandlers';
-import { logout } from './utils/handlers/authenticationHandlers';
+import withRestrictedAccess from '../hoc/withRestrictedAccess';
+import Dashboard from '../pages/Dashboard/Dashboard';
+import Login from '../pages/Login/Login';
+import CreateAccount from '../pages/CreateAccount/CreateAccount';
+import { getUser } from '../../utils/handlers/authenticationHandlers';
+import { logout } from '../../utils/handlers/authenticationHandlers';
 
 function App() {
   const [user, setUser] = useState(undefined);
