+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App/App';
 import 'semantic-ui-css/semantic.min.css';
 import * as serviceWorker from './utils/serviceWorker';
